Add better demosaic
Former-commit-id: dc16ec30236455eccf68615786fd2d18ee04f9d5 [formerly 9807bbf987ff5439054c2dd4bf848584d617be81]
Former-commit-id: ace9674d5ea492bac72de47a69492186683397d5 [formerly c193a4b6c78913904ad9ce2d5ec539c2a4baceaf]
Former-commit-id: be9c272ad97ae2af465997db2a1022988f544120
Former-commit-id: 73a2eb7becc12bcb4e42218a79f490bed67f2d66

+++ Source/RawParser/Model/Parser/Decoder/TIFFDecoder.cs
@@ -309,7 +309,7 @@ namespace RawNet.Decoder
                         longitudeRef = gps.GetEntry((TagType)0x02)?.DataAsString,
                         lattitudeRef = gps.GetEntry((TagType)0x02)?.DataAsString,
                         altitude = gps.GetEntry((TagType)0x02)?.GetFloat(0) ?? 0,
-                        altitudeRef = gps.GetEntry((TagType)0x02)?.GetInt(0) ?? 0
+                        altitudeRef = gps.GetEntry((TagType)0x02)?.GetLong(0) ?? 0
                     };
                 }
             }
